Added checking for no value given
This have worked in most cases, however
if we are expecting a string the checking
did not work because 'unset' is a string.

Now we are explicitly checking if the value
is an 'unset' string, and if that is the case
the type-checker will throw an error.

It might be a good idea to change the 'unset'
from a string to something else in the future
in order to avoid similar issues (and the user
might want to send an 'unset' string to the function
and we are currently preventing that from being possible).

Done as part of issue #15