Changed 'unset' and 'pass' to obj
Internally the type-checker now uses
unset and ignore objects instead of
strings. This is done to make handling
of default values easier and also to
avoid possible issues with string types.

The user can still manually set a parameter
to be ignored using the 'pass' string.

Done as part of issue #10